allow init steps (match behave functionality)
